# profileService/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: Profile Service API
  description: |
    API сервиса управления профилями преподавателей и курсами.
    Предоставляет функционал для:
    - Управления профилями преподавателей
    - Управления курсами и их экземплярами
    - Фильтрации данных
    - Распределения преподавателей по курсам
    - Парсинга Excel файлов
  version: 1.0.0
  contact:
    name: FAH Team
    email: support@innopolis.university
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.fah.innopolis.university
    description: Production server
...
paths:
...
  /profile/addProfile:
    post:
      tags:
        - Profile
      summary: Добавление нового профиля преподавателя
      description: Создаёт новый профиль преподавателя с указанными параметрами
      operationId: addProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProfileRequest'
      responses:
        '200':
          description: Профиль успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddProfileResponse'
        '400':
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/getProfile/{id}:
    get:
      tags:
        - Profile
      summary: Получение профиля по ID
      description: Возвращает полную информацию о профиле преподавателя по версии профиля
      operationId: getProfile
      parameters:
        - name: id
          in: path
          required: true
          description: ID версии профиля
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешное получение профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfileResponse'
        '400':
          description: Некорректный ID профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/getAllProfiles:
    get:
      tags:
        - Profile
      summary: Получение списка всех профилей
      description: Возвращает список профилей преподавателей с возможностью фильтрации по институту и должности
      operationId: getAllProfiles
      parameters:
        - name: year
          in: query
          required: true
          description: Год
          schema:
            type: integer
            format: int64
        - name: institute
          in: query
          required: false
          description: ID института (можно передать несколько)
          schema:
            type: array
            items:
              type: integer
              format: int64
        - name: position
          in: query
          required: false
          description: ID должности (можно передать несколько)
          schema:
            type: array
            items:
              type: integer
              format: int64
      responses:
        '200':
          description: Успешное получение списка профилей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllFacultiesResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== COURSES ====================
...
  /filter/profile:
    get:
      tags:
        - Filters
      summary: Получение фильтров для профилей
      description: Возвращает доступные значения фильтров для списка профилей (институты и должности)
      operationId: getFacultyFilters
      responses:
        '200':
          description: Успешное получение фильтров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFacultyFiltersResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
      required:
        - error

    # ==================== PROFILE SCHEMAS ====================
    AddProfileRequest:
      type: object
      properties:
        name_eng:
          type: string
          description: Имя на английском языке
        email:
          type: string
          format: email
          description: Email преподавателя
        alias:
          type: string
          description: Псевдоним преподавателя
        institute_id:
          type: array
          items:
            type: integer
            format: int64
          description: Список ID институтов
        position_id:
          type: integer
          format: int64
          description: ID должности
        year:
          type: integer
          format: int64
          description: Год
      required:
        - name_eng
        - institute_id
        - position_id
        - year
    AddProfileResponse:
      type: object
      properties:
        profile_id:
          type: integer
          format: int64
          description: ID версии профиля
        name_eng:
          type: string
          description: Имя на английском языке
        email:
          type: string
          description: Email
        alias:
          type: string
          description: Псевдоним
        institute_names:
          type: array
          items:
            type: string
          description: Названия институтов
        position_name:
          type: string
          description: Название должности
        year:
          type: integer
          format: int64
          description: Год
    GetProfileResponse:
      type: object
      properties:
        profile_id:
          type: integer
          format: int64
          description: ID версии профиля
        year:
          type: integer
          format: int64
          description: Год
        name_eng:
          type: string
          description: Имя на английском языке
        name_ru:
          type: string
          nullable: true
          description: Имя на русском языке
        alias:
          type: string
          description: Псевдоним
        email:
          type: string
          description: Email
        position_name:
          type: string
          description: Название должности
        institute_names:
          type: array
          items:
            type: string
          description: Названия институтов
        student_type:
          type: string
          nullable: true
          description: Тип студента
        degree:
          type: boolean
          nullable: true
          description: Наличие степени
        fsro:
          type: string
          nullable: true
          description: FSRO
        languages:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Language'
          description: Языки
        employment_type:
          type: string
          nullable: true
          description: Тип занятости
        hiring_status:
          type: string
          nullable: true
          description: Статус найма
        mode:
          type: string
          nullable: true
          description: Режим работы
        max_load:
          type: integer
          format: int64
          nullable: true
          description: Максимальная нагрузка
        frontal_hours:
          type: integer
          format: int64
          nullable: true
          description: Фронтальные часы
        extra_activities:
          type: number
          format: double
          nullable: true
          description: Дополнительные активности
        workload_stats:
          $ref: '#/components/schemas/WorkloadStats'
          nullable: true
          description: Статистика нагрузки
    GetAllFacultiesResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/ShortProfile'
          description: Список профилей
    ShortProfile:
      type: object
      properties:
        profile_id:
          type: integer
          format: int64
          description: ID версии профиля
        name_eng:
          type: string
          description: Имя на английском языке
        alias:
          type: string
          description: Псевдоним
        email:
          type: string
          description: Email
        position_name:
          type: string
          description: Название должности
        institute_names:
          type: array
          items:
            type: string
          description: Названия институтов
    Language:
      type: object
      properties:
        language_code:
          type: string
          description: Код языка
    WorkloadStats:
      type: object
      properties:
        t1:
          $ref: '#/components/schemas/Classes'
          description: Статистика за триместр 1
        t2:
          $ref: '#/components/schemas/Classes'
          description: Статистика за триместр 2
        t3:
          $ref: '#/components/schemas/Classes'
          description: Статистика за триместр 3
        total:
          $ref: '#/components/schemas/Classes'
          description: Общая статистика
    Classes:
      type: object
      properties:
        lec_hours:
          type: integer
          format: int64
          description: Часы лекций
        tut_hours:
          type: integer
          format: int64
          description: Часы тьюториалов
        lab_hours:
          type: integer
          format: int64
          description: Часы лабораторных
        elective_hours:
          type: integer
          format: int64
          description: Часы элективов
        rate:
          type: number
          format: double
          description: Ставка

    # ==================== COURSE SCHEMAS ====================
...
    FilterObj:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID объекта фильтра
        name:
          type: string
          description: Название объекта фильтра
    GetFacultyFiltersResponse:
      type: object
      properties:
        institute:
          type: array
          items:
            $ref: '#/components/schemas/FilterObj'
          description: Список институтов для фильтрации
        position:
          type: array
          items:
            $ref: '#/components/schemas/FilterObj'
          description: Список должностей для фильтрации
...